feat(Drawer): style Drawer to expand dynamically
See the 'flex' property at
https://facebook.github.io/react-native/docs/layout-props.html :

In React Native flex does not work the same way that it does in CSS. flex is a
number rather than a string, and it works according to the css-layout library
at https://github.com/facebook/css-layout.

When flex is a positive number, it makes the component flexible and it will be
sized proportional to its flex value. So a component with flex set to 2 will
take twice the space as a component with flex set to 1.

When flex is 0, the component is sized according to width and height and it is
inflexible.

When flex is -1, the component is normally sized according width and height.
However, if there's not enough space, the component will shrink to its minWidth
and minHeight.