feat(SplashContainer): dispatch handleAuthWithFirebase
In SplashContainer, we want to authenticate with firebase after authentication
with Facebook and retrieving an access token.

Also in SplashContainer, we don't want to mapStateToProps or anything.
We need react-redux to connect the component, thereby passing "dispatch" as a
prop, which we can then use to dispatch handleAuthWithFirebase.

So in summary: Splash view is shown, the user authenticates with Facebook, once
they authenticate, onLoginFinished(== handleLoginFinished) is invoked. If
everything is well, this will dispatch handleAuthWithFirebase(), which
dispatches authenticating which sets isAuthenticating to true, and calls
getAccessToken and authWithToken.

BUT when successful, the state will still have isAuthed == false and isAuthenticating == true.