feat(navbar): hook up Settings to the Gear button
We pass a handleToSettings function to the Home component which
pushes the settings route to the navigator.

On the initial run, route is undefined, so we must set Navigator's
initialRoute prop to something (here, just an empty object),
toherwise ReactModoroNavigator would yield

> Possible Unhandled Promise Rejection (id: 0):
> Cannot read property 'settings' of undefined.

We also need to setup a scene configuration for settings, otherwise

> Cannot read property 'defaultTransitionVelocity' of undefined

is thrown.